requirements.txt
pandas
openpyxl
python-dotenv
sqlalchemy
psycopg2
pymssql